implement ABCs and type hint them
