+++ pyproject.toml
@@ -26,4 +26,5 @@ precision = 1
 exclude_lines = [
   "pragma: no cover",
   "abc.abstractmethod",
+  "abstractmethod",
   ]
